refactor: User package 재구성
