+++ backend/src/main/java/anyone/to/soma/user/domain/User.java
@@ -1,6 +1,6 @@
 package anyone.to.soma.user.domain;
 
-import anyone.to.soma.user.LoginType;
+import anyone.to.soma.user.domain.type.LoginType;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -62,7 +62,7 @@ public class User {
     }
 
     public String getNickname() {
-        if(profile == null) {
+        if (profile == null) {
             return "익명";
         }
